+++ src/app/(Product)/wedding/invite/[id]/_components/InviteCalender/hooks/useConvertInviteDate/useConvertInviteDate.type.ts
@@ -1,0 +1,3 @@
+export interface UseConvertInviteDate {
+  calenderData: string;
+}
